Update for Develocity
