+++ .github/workflows/convention-develocity-maven-extension-verification.yml
@@ -1,0 +1,57 @@
+name: Verify Convention Develocity Maven Extension
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+  workflow_dispatch:
+
+jobs:
+  build:
+    name: Build Convention Develocity Maven Extension
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Set up JDK 8
+        uses: actions/setup-java@v4
+        with:
+          java-version: '8'
+          distribution: 'temurin'
+          cache: maven
+      - name: Build with Maven
+        working-directory: convention-develocity-maven-extension/extension
+        run: ./mvnw clean install
+      - name: Upload extension
+        uses: actions/upload-artifact@v4
+        with:
+          name: convention-develocity-maven-extension
+          path: ~/.m2/repository/com/myorg
+
+  verification:
+    name: Verify Example Build for Maven ${{ matrix.maven-version }}
+    needs: [build]
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        maven-version: ['3.6.3', '3.8.8', '3.9.1']
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Set up JDK 8
+        uses: actions/setup-java@v4
+        with:
+          java-version: '8'
+          distribution: 'temurin'
+          cache: maven
+      - name: Download extension
+        uses: actions/download-artifact@v4
+        with:
+          name: convention-develocity-maven-extension
+          path: ~/.m2/repository/com/myorg
+      - name: Verify example build
+        working-directory: convention-develocity-maven-extension/example
+        run: ./mvnw clean verify -Ddevelocity.url=https://ge.solutions-team.gradle.com
+        env:
+          DEVELOCITY_ACCESS_KEY: ${{ secrets.GE_SOLUTIONS_ACCESS_TOKEN }}
