--- .github/workflows/convention-gradle-enterprise-maven-extension.yml
@@ -1,63 +1,0 @@
-name: Verify Maven Convention Extension
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-  workflow_dispatch:
-
-jobs:
-  build:
-    name: Build Gradle Enterprise Conventions Maven Extension
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Set up JDK 8
-        uses: actions/setup-java@v3
-        with:
-          java-version: '8'
-          distribution: 'temurin'
-          cache: maven
-      - name: Build Gradle Enterprise Conventions Maven Extension
-        run: ./mvnw clean install -Dgradle.enterprise.url=https://ge.solutions-team.gradle.com
-        working-directory: convention-gradle-enterprise-maven-extension/extension
-        env:
-          GRADLE_ENTERPRISE_ACCESS_KEY: ${{ secrets.GE_SOLUTIONS_ACCESS_TOKEN }}
-      - name: Upload extension
-        uses: actions/upload-artifact@v3
-        with:
-          name: convention-gradle-enterprise-maven-extension
-          path: ~/.m2/repository/com/myorg/convention-gradle-enterprise-maven-extension
-  verification:
-    name: Verify Example Build for Maven ${{ matrix.maven-version }}
-    needs: [build]
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        maven-version: ['3.6.3', '3.8.8', '3.9.1']
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Set up JDK 8
-        uses: actions/setup-java@v3
-        with:
-          java-version: '8'
-          distribution: 'temurin'
-          cache: maven
-      - name: Download Extension
-        uses: actions/download-artifact@v3
-        with:
-          name: convention-gradle-enterprise-maven-extension
-          path: ~/.m2/repository/com/myorg/convention-gradle-enterprise-maven-extension
-      - name: Set Maven Version
-        run: ./mvnw wrapper:wrapper -Dmaven=${{ matrix.maven-version }} -Dgradle.enterprise.url=https://ge.solutions-team.gradle.com
-        working-directory: convention-gradle-enterprise-maven-extension/example
-        env:
-          GRADLE_ENTERPRISE_ACCESS_KEY: ${{ secrets.GE_SOLUTIONS_ACCESS_TOKEN }}
-      - name: Verify Example Build
-        run: ./mvnw clean verify -Dgradle.enterprise.url=https://ge.solutions-team.gradle.com
-        working-directory: convention-gradle-enterprise-maven-extension/example
-        env:
-          GRADLE_ENTERPRISE_ACCESS_KEY: ${{ secrets.GE_SOLUTIONS_ACCESS_TOKEN }}
