+++ convention-develocity-maven-extension/.gitignore
@@ -1,0 +1,3 @@
+target/
+dependency-reduced-pom.xml
+**/.mvn/.develocity/develocity-workspace-id
