+++ convention-develocity-maven-extension/example/.mvn/extensions.xml
@@ -3,7 +3,7 @@
             xsi:schemaLocation="http://maven.apache.org/EXTENSIONS/1.0.0 http://maven.apache.org/xsd/core-extensions-1.0.0.xsd">
     <extension>
         <groupId>com.myorg</groupId>
-        <artifactId>convention-gradle-enterprise-maven-extension</artifactId>
+        <artifactId>convention-develocity-maven-extension</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </extension>
 </extensions>
