+++ convention-develocity-maven-extension/example/pom.xml
@@ -6,7 +6,7 @@
     <packaging>jar</packaging>
     <version>1.0</version>
 
-    <name>Convention Gradle Enterprise Maven Extension Example</name>
+    <name>Convention Develocity Maven Extension Example</name>
 
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
