+++ convention-develocity-maven-extension/example/src/main/java/com/myorg/Example.java
