--- convention-gradle-enterprise-maven-extension/.gitignore
@@ -1,3 +1,0 @@
-target/
-dependency-reduced-pom.xml
-**/.mvn/.gradle-enterprise/gradle-enterprise-workspace-id
