--- convention-gradle-enterprise-maven-extension/example/.gitattributes
@@ -1,1 +1,0 @@
-*.bat           text eol=crlf
