--- convention-gradle-enterprise-maven-extension/extension/.gitattributes
@@ -1,1 +1,0 @@
-*.bat           text eol=crlf
